Prototype 0.4
Added code to gather input from touch sensor on the arduino and print
to serial (currently commented out). Added a countdown timer to the
processing sketch and tweaked some temperature variables. Still need to
parse through serial line to separate true false of the button from the
float of the temperature reading.